requirements.txt
Django==2.1.7
django-bootstrap4==0.0.7
djangorestframework==3.9.2
pkg-resources==0.0.0
pytz==2018.9
RPi.GPIO==0.6.5
